# app/api/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: ROSSO API
  description: API для управления поставками ROSSO
  version: 1.0.0


paths:
  /auth/register:
    post:
      summary: Регистрация пользователя
      tags:
        - Аутентификация и регистрация
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        "201":
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  /auth/login:
    post:
      summary: Авторизация пользователя
      tags:
        - Аутентификация и регистрация
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        "200":
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  /users/{user_id}:
    get:
      summary: Получение пользователя
      tags:
        - Пользователи
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSchema'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
    put:
      summary: Изменение данных пользователя
      tags:
        - Пользователи
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSchema'
      responses:
        "200":
          description: ОК
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSchema'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
        
  /user/company/{link_code}:
    get:
      summary: Получение пользователя по пригласительному коду
      tags:
        - Пользователи
      operationId: getUserByLinkCode
      parameters:
        - name: link_code
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ОК
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name: 
                    type: string
                  email: 
                    type: string
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
    
    delete:
      summary: Привязка пользователя к уч. записи компании
      tags:
        - Пользователи
      operationId: deleteUserByLinkCode
      parameters:
        - name: link_code
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: ОК
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  
  /user/company:
    post:
      summary: Привязка пользователя к уч. записи компании
      tags:
        - Пользователи
      operationId: postUserByLinkCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                id_company:
                  type: integer
                role:
                  type: string
                  enum: ["manager", "employee"]
      responses:
        '201':
          description: ОК
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    
                    
  /organizers/register:
    post:
      summary: Регистрация организации
      tags:
        - Организации
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizerRegisterRequest'
      responses:
        "201":
          description: Организация успешно зарегистрирована
        "422":
          $ref: '#/components/responses/UnprocessableEntity'


  /organizers:
    get:
      summary: Получение организаций, с которыми компания заключила контракт
      tags:
        - Организации
      responses:
        "200":
          description: Список организаций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizersResponse'

  /products:
    post:
      summary: Добавить новый товар
      tags:
        - Товары
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        "201":
          description: Товар успешно добавлен
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    get:
      summary: Получить список товаров
      tags:
        - Товары
      responses:
        "200":
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsResponse'

  /products/{product_id}:
    get:
      summary: Получить информацию о товаре
      tags:
        - Товары
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Информация о товаре
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    put:
      summary: Обновить информацию о товаре
      tags:
        - Товары
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductResponse'
      responses:
        "200":
          description: Товар успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'


  /supplies:
    get:
      summary: Получить список всех поставок
      tags:
        - Поставки
      responses:
        "200":
          description: Список поставок
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuppliesResponse'
                
    post:
      summary: Создать новую поставку
      tags:
        - Поставки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyCreateRequest'
      responses:
        "201":
          description: Поставка успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyResponse'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  /supplies/{supply_id}/status:
    put:
      summary: Изменить статус поставки
      tags:
        - Поставки
      parameters:
        - name: supply_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyStatusUpdate'
      responses:
        "200":
          description: Статус поставки успешно изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyResponse'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  /suppliers:
    get:
      summary: Получить список поставщиков (фильтр - название)
      tags:
        - Поставщики
      responses:
        "200":
          description: Список поставщиков
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuppliersResponse'

  /suppliers/{supplier_id}:
    post:
      summary: Добавить нового поставщика в контакт
      tags:
        - Поставщики
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Поставщик успешно добавлен
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      summary: Удалить поставщика из контактов
      tags:
        - Поставщики
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Поставщик успешно удалён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerResponse" 
        "404":
          $ref: '#/components/responses/NotFound'


  /suppliers/products/{supplier_id}:
    get:
      summary: Получить список продуктов поставщика
      tags:
        - Поставщики
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Список продуктов поставщика
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsResponse'
        "404":
          $ref: '#/components/responses/NotFound'

  /expenses/:
    get:
      summary: Получить все расходы
      tags:
        - Расходы склада
      operationId: getExpenses
      responses:
        '200':
          description: Список всех расходов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expenses'
    post:
      summary: Добавить новый расход
      tags:
        - Расходы склада
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        '201':
          description: Расход добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'

  /expenses/{expense_id}/:
    patch:
      summary: Обновить количество расхода
      tags:
        - Расходы склада
      operationId: updateExpenseQuantity
      parameters:
        - name: expense_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
      responses:
        '200':
          description: Количество обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'

    delete:
      summary: Удалить расход
      tags:
        - Расходы склада
      operationId: deleteExpense
      parameters:
        - name: expense_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Расход удален

components:
  responses:
    NotFound:
      description: not found resource response
      content:
        application/JSON:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: payload value is invalid
      content:
        application/JSON:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        details:
          oneOf:
            - type: string
            - type: array
      required:
        - details

    # Пользовательские схемы
    UserRegisterRequest:
      type: object
      properties:
        role:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string

    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        type_token:
          type: string

    UserSchema:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string

    # Схемы организаций
    OrganizerRegisterRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [ company, supplier ]
        address:
          type: string
        inn:
          type: string
        bank_details:
          type: string

    OrganizerResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [ company, supplier ]
        address:
          type: string
        inn:
          type: string
        bank_details:
          type: string

    OrganizersResponse:
      type: object
      properties:
        organizers:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              role:
                type: string
                enum: [ company, supplier ]
              address:
                type: string
              inn:
                type: string
              bank_details:
                type: string

    
    # Схемы поставок
    SupplyCreateRequest:
      type: object
      properties:
        supplier_id:
          type: integer
        delivery_address:
          type: string
        total_price:
          type: number
          format: float
        supply_products:
          type: array
          items:
            type: object
            properties:
              product:
                $ref: "#/components/schemas/ProductResponseSupply"
              quantity:
                type: integer


    SupplyResponse:
      type: object
      properties:
        supplier:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
        supply_products:
          type: array
          items:
            type: object
            properties:
              product:
                $ref: "#/components/schemas/ProductResponseSupply"
              quantity:
                type: integer
        couriers_phone:
          type: string
        article:
          type: integer
        supplier_id:
          type: integer
        delivery_address:
          type: string
        total_price:
          type: number
          format: float
        status:
          type: string
          enum: [in_processing, assembled, in_delivery, adopted]
        create_datetime: 
          type: string
        delivery_datetime:
          type: string

    SuppliesResponse:
      type: object
      properties:
        supplies:
          type: array
          items:
            $ref: '#/components/schemas/SupplyResponse'

    SupplyStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [in_processing, assembled, in_delivery, adopted]

    # Схемы товаров в поставках
    SuppliersResponse:
      type: object
      properties:
        organizers:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              address:
                type: string
              inn:
                type: string
              bank_details:
                type: string


    # Схемы товаров
    ProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [ 
            hair_coloring, 
            hair_care, 
            hair_styling, 
            consumables, perming, 
            eyebrows, 
            manicure_and_pedicure, 
            tools_and_equipment 
            ]
        price:
          type: number
          format: float

    ProductResponse:
      type: object
      properties:
        id:
          type: integer
        article:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
          format: float

    ProductResponseSupply:
      type: object
      properties:
        id:
          type: integer
        article:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [ 
            hair_coloring, 
            hair_care, 
            hair_styling, 
            consumables, perming, 
            eyebrows, 
            manicure_and_pedicure, 
            tools_and_equipment 
            ]
        price:
          type: number
          format: float

    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponseSupply'

    Expenses:
      type: object
      properties:
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
              
    Expense:
      type: object
      required:
        - company_id
        - article
        - name
        - category
        - quantity
      properties:
        company_id:
          type: integer
        article:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [hair_coloring, hair_care, hair_styling, consumables, perming, eyebrows_and_eyelashes, manicure_and_pedicure, tools_and_equipment]
        quantity:
          type: integer
    
    ExpenseCreate:
      type: object
      required:
        - company_id
        - article
        - name
        - category
        - quantity
      properties:
        company_id:
          type: integer
        article:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer