# app/schemas/supply.py
from typing import Optional, List
from pydantic import BaseModel, ConfigDict

from organizer import OrganizerSupplyObject
from product import ProductResponseSupply

class SupplyBase(BaseModel):
    supplier_id: int
    delivery_address: str
    total_price: float


class SupplyProduct(BaseModel):
    product: ProductResponseSupply
    quantity: int


class SupplyCreateRequest(SupplyBase):
    supply_products: List[SupplyProduct]


class SupplyResponse(SupplyBase):
    supplier: OrganizerSupplyObject
    supply_products: List[ProductResponseSupply]
    couriers_phone: str
    article: int
    status: str
    create_datetime: str
    delivery_datetime: Optional[str]


class SuppliesResponse(BaseModel):
    supplies: List[SupplyResponse]


class SupplyStatusUpdate(BaseModel):
    status: str